Type-disagreement repair: 104 fixed, 255 provably-missing args found, flash delta 0
Stage 03's module headers put caller and callee in one TU and surfaced 1,014 type
disagreements. This pass arbitrates every one of them against the shipped image.

TWO NEW IMAGE ORACLES (the reusable output)
scratchpad/td/callee.py: abstract interpretation over each function's CFG giving
in_regs (r0-r3 read-before-written = an EXACT LOWER BOUND ON ARITY, needing no
caller and no format string), in_stack, in_vfp, ret_r0_expl/ret_r1_expl. Swept
the whole app image in 3 s: 2,377 functions + 13,713 call sites. TWO CALIBRATION
BUGS found and fixed while validating -- interleaved GCC prologues broke frame
math, and Ghidra data-inflation polluted stack-arg detection. Validated against
four independently-known prototypes.

FOUR-WAY SPLIT (FIXED is measured by RE-RUNNING Stage 03's own agreement test,
not by counting edits)
  class                                FIXED  CONFIRMED  FALSE ALARM  UNSETTLED
  arity_differs                            0        117            0        157
  declaration_disagrees_with_definition   62         41          141        202
  parameter_type_differs                  28         28           49         14
  return_type_differs                     14         31           73         57
  TOTAL                                  104        217          263        430

KEY FINDINGS
- 255 CALL SITES PASS FEWER ARGUMENTS THAN THE CALLEE'S PROLOGUE PROVABLY READS,
  by register-only proof with no heuristics. This is the register-passthrough
  shape (strlen(), k_sleep() after __aeabi_uldivmod) that Ghidra cannot express
  and cfg_verify is blind to. 53 repaired under a strictly mechanical rule (all
  missing regs trace to function entry; unambiguous source<->machine match) --
  and FLASH MOVED 0 BYTES, because the arguments were already in the registers.
- 6 DEFINITIONS declare fewer parameters than the image proves they read
  (lc3_sns_analyze 7 vs 13, z_impl_k_timer_start 3 vs 6, ...). Previously invisible.
- __ieee754_exp.c LOCALLY REDEFINES u8 AS uint64_t while g1_types.h says uint8_t.
  Merging lib/ into a cohesive TU would SILENTLY TRUNCATE every soft-double
  helper return to 8 bits -- not a compile error. Highest-risk item found;
  recorded, not fixed.
- R8 triage of the 303 invented-argument sites: 287 inert / 21 volatile / 0 call.
- Narrow-vs-wide parameter spelling is NOT cosmetic: measured uxtb emission at
  the call.

ACCEPTANCE -- all green
build exit 0; nm -u 0; 0 dup globals; pins 627/0/0;
check_thread_create_stack_args 10/10; verify_data 995/995;
FLASH 956,480 B / 97.35 % -- DELTA 0, headroom 26,048 B; RAM delta 0; net
untouched; per-object 2,827/2,837 identical.
ALL FOUR framebuffers byte-identical by cmp under a SEEDED capture. Iteration-41
wins re-measured and held: spim_a 3,645; twim2 p1_boot 1,089 / 7ed8ddcd0c0d... EQ;
whole twim1 p1_boot per-device set EQ; JBD_FRAMECOUNTER_P2 0x0D61 EQ.
(ESB_MASTER_FRAMES 0x176 vs 0x175 -- inside the oracle's documented +/-2
radio-cadence tolerance.)

WITHDRAWN, AND WHY -- this is the important part
A relaxed Pass 1b resolved 213 entries instead of 104, built clean, and kept all
four framebuffers -- but its +60 B of .text REORDERED TWO FIRMWARE THREADS ON THE
SHARED I2C BUS, breaking twim2 p1_boot and opt3001 byte-equality. Diffing with
ticks stripped shows a pure PERMUTATION of an identical transaction multiset, so
it is very likely benign -- but "probably benign timing" is not this project's
standard, so it was FULLY REVERTED. 109 resolvable entries given up; canon2.py
kept for a future pass with its own baseline.
Pass 1 also regressed check_thread_create_stack_args to 9/10 via
dev_api_call_slot0; that symbol was withdrawn and the gate is 10/10.

NOT CLOSED (10 items in 9): 910 entries remain; 202 of the 255 missing-argument
sites (a WRONG argument is worse than a missing one); the 6 short definitions;
the 303 invented arguments (deleting ~287 setups moves .text far more than the
60 B that just broke a gate -- needs its own pass and baseline); the 36
key_event_thread sites (no oracle applies); the u8 typedef collision.

Detail: recon/analysis/type_disagreement_repair.md (802 lines).

+++ recon/app/src/FUN_000125c0.c
@@ -1,7 +1,7 @@
 /* Reconstructed FUN_000125c0 @ 0x125c0  (parity: 300/300 trials, PROVEN) */
 
 #include <stdint.h>
-extern uint64_t FUN_0000d89c(int32_t, int32_t);
+extern unsigned long long FUN_0000d89c(unsigned int, unsigned int);
 extern uint64_t FUN_0000d8f8(uint32_t, uint32_t, uint32_t, uint32_t);
 extern int64_t FUN_0000ec34(uint32_t, uint32_t);
