+++ recon/app/src/FUN_0004d6ec.c
@@ -6,7 +6,7 @@
 extern void FUN_0007e2fa(unsigned long, ...);
 extern void FUN_0007e2ec(uint32_t, uint32_t) __attribute__((noreturn));
 extern uintptr_t FUN_0004d334(uint32_t, uint32_t);
-extern uintptr_t FUN_0004d2d0(uintptr_t);
+extern unsigned int FUN_0004d2d0(unsigned int);
 extern uintptr_t FUN_000748ac(void);
 extern void FUN_00072880(uintptr_t);
 extern uint64_t FUN_0004d594(void);
